requirements.txt
authy==2.2.3
Flask==1.0.2
phonenumbers==8.10.1
python-dotenv==0.9.1
